setup.py: remove shell script install, install.sh will handle this, update readme.md, move macos python scripts into fetch_cord folder.
